clean up, add some comments
